chore(backends): convert scalars to non-numpy python objects (#10233)
This PR converts scalar numpy objects into their non-numpy Python
equivalents for consistent scalar return types for all backends.

BREAKING CHANGE: `execute` now returns non-numpy objects for scalar values.

+++ ibis/expr/api.py
@@ -199,11 +199,11 @@ def param(type: dt.DataType) -> ir.Scalar:
     ... )
     >>> expr = t.filter(t.date_col >= start).value.sum()
     >>> expr.execute(params={start: date(2013, 1, 1)})
-    np.float64(6.0)
+    6.0
     >>> expr.execute(params={start: date(2013, 1, 2)})
-    np.float64(5.0)
+    5.0
     >>> expr.execute(params={start: date(2013, 1, 3)})
-    np.float64(3.0)
+    3.0
     """
     return ops.ScalarParameter(type).to_expr()
